Add adjustment rules based on percentile

+++ EvoBio3.Core/Extensions/ListExtensions.cs
@@ -34,7 +34,7 @@ namespace EvoBio3.Core.Extensions
 
 		[Pure]
 		public static T AtPercentage<T> ( this IList<T> list,
-		                                  int percentage ) =>
+		                                  double percentage ) =>
 			list[(int) Math.Round ( ( list.Count - 1d ) * percentage / 100d )];
 	}
 }
